Extract common object prim code
